ExpiryTracker. Clean up Detail UI refreshes with Food updates
Rationale: Whenever a food gets updated, DetailActivity would recreate
all the fragments in its Adapter. On top of this, each fragment was
already handling its own food updates. While this is happening, the
user sees a jarring refresh effect, which doesn't look nice at all.
The solution is to let each fragment continue to update the UI with
updated info, while only updating the Adapter when a food gets removed.
We add a ProgressBar for each fragment to make the data loads look nicer

Signed-off-by: trod-123 <cia.123trod@gmail.com>