ExpiryTracker. Tweak List UI further
- Do not let RTD listeners refresh list when pushing a recent change
- Do not let RTD listeners recreate food after images are uploaded to FS
- Fix bug where RTD refresh may query a -1 index in FoodAdapter

Rationale: Concerning the first bullet, to make this work we had to also
update the local timestamp. Originally this was not done - only the
remote timestamp was updated, pushing a redundant local update, causing
wonky UI refreshes. However, we did not want to use a System generated
timestamp because users may access their data in different timezones.
The solution for getting a locale-neutral timestamp was to use
Firebase RTD's generated key value via push(). We then store this
"timestamp" locally, and write it remotely during updates, taking the
place of the existing Long timestamp

Concerning the second bullet, when Firebase Storage replaces local
image uris with Storage uris for a food item, RTD listeners think this
operation creates a new food item, even though fields of that food item
are being updated. This is due to the nature of the
FirebaseReference.setValue() method, which triggers the onChildAdded()
call in the ChildEventListener. This causes the UI to very briefly
delete and re-insert the affected food item, which isn't nice. The
solution for this was to first decouple the RTD write process by first
writing the food item to RTD with the local image uris, using
ref.setValue() as usual. Afterwards, we upload the local images to
Storage. Once we get all the Storage uris, instead of rewriting the food
item in RTD with ref.setValue(), we call an update to the food item
with ref.updateChildren(). This ensures the other food fields are
untouched, which in turn triggers onChildChanged() in
ChildEventListener. This allows the affected food item to NOT be
re-created, and instead refreshes just the images. Much cleaner

Signed-off-by: Teddy Rodriguez <cia.123trod@gmail.com>