ExpiryTracker. Modify legal urls to static Firebase Storage refs
Rationale: Download urls to files in Firebase Storage change even when
the file is updated, so storing the download url in Java is not a good
idea when Storage files may potentially be updated. Instead, store the
Firebase Storage uri in Java, and then grab the download url via
the Storage uri. Just make sure you don't change the filename or path
in Firebase Storage.

Signed-off-by: Teddy Rodriguez <cia.123trod@gmail.com>